Refactor: Move bluetooth helpers from AppContext into BluetoothHelpers
