+++ app/src/main/java/de/simon/dankelmann/bluetoothlespam/AppContext/AppContext.kt
@@ -15,20 +15,6 @@ abstract class AppContext {
         private lateinit var _bluetoothLeScanService: IBluetoothLeScanService
         private lateinit var _advertisementSetQueueHandler: AdvertisementSetQueueHandler
 
-        fun Context.bluetoothAdapter(): BluetoothAdapter? = (this.getSystemService(Context.BLUETOOTH_SERVICE) as BluetoothManager).adapter
-        fun Context.bluetoothManager(): BluetoothManager? = (this.getSystemService(Context.BLUETOOTH_SERVICE) as BluetoothManager)
-
-        fun isBluetooth5Supported():Boolean{
-            val bluetoothAdapter:BluetoothAdapter? = this.getContext().bluetoothAdapter()
-            if(bluetoothAdapter != null) {
-                return (bluetoothAdapter.isLe2MPhySupported
-                        && bluetoothAdapter.isLeCodedPhySupported
-                        && bluetoothAdapter.isLeExtendedAdvertisingSupported
-                        && bluetoothAdapter.isLePeriodicAdvertisingSupported)
-            }
-            return false
-        }
-
         fun setContext(context: Context) {
             _context = context
         }
