+++ app/src/main/java/de/simon/dankelmann/bluetoothlespam/Helpers/BluetoothHelpers.kt
@@ -1,9 +1,10 @@
 package de.simon.dankelmann.bluetoothlespam.Helpers
 
+import android.bluetooth.BluetoothAdapter
+import android.bluetooth.BluetoothManager
 import android.content.Context
 import androidx.preference.PreferenceManager
 import de.simon.dankelmann.bluetoothlespam.AppContext.AppContext
-import de.simon.dankelmann.bluetoothlespam.AppContext.AppContext.Companion.bluetoothAdapter
 import de.simon.dankelmann.bluetoothlespam.Interfaces.Services.IAdvertisementService
 import de.simon.dankelmann.bluetoothlespam.Interfaces.Services.IBluetoothLeScanService
 import de.simon.dankelmann.bluetoothlespam.R
@@ -12,19 +13,20 @@ import de.simon.dankelmann.bluetoothlespam.Services.LegacyAdvertisementService
 import de.simon.dankelmann.bluetoothlespam.Services.ModernAdvertisementService
 
 class BluetoothHelpers {
+
     companion object {
-        fun supportsBluetooth5(): Boolean {
-            var bluetoothAdapter = AppContext.getContext().bluetoothAdapter()
-            if (bluetoothAdapter != null) {
-                if (bluetoothAdapter!!.isLe2MPhySupported
-                    && bluetoothAdapter!!.isLeCodedPhySupported
-                    && bluetoothAdapter!!.isLeExtendedAdvertisingSupported
-                    && bluetoothAdapter!!.isLePeriodicAdvertisingSupported
-                ) {
-                    return true
-                }
-            }
-            return false
+
+        fun Context.bluetoothManager(): BluetoothManager? =
+            (this.getSystemService(Context.BLUETOOTH_SERVICE) as BluetoothManager)
+
+        fun Context.bluetoothAdapter(): BluetoothAdapter? = this.bluetoothManager()?.adapter
+
+        fun Context.isBluetooth5Supported(): Boolean {
+            val bluetoothAdapter = this.bluetoothAdapter() ?: return false
+            return (bluetoothAdapter.isLe2MPhySupported
+                    && bluetoothAdapter.isLeCodedPhySupported
+                    && bluetoothAdapter.isLeExtendedAdvertisingSupported
+                    && bluetoothAdapter.isLePeriodicAdvertisingSupported)
         }
 
         fun getAdvertisementService(context: Context): IAdvertisementService {
